Suport files larger than 2Gb. Fixes issue 171. Fixes issue 173

+++ core/src/main/java/gwtupload/server/UploadListener.java
@@ -134,7 +134,7 @@ public class UploadListener extends AbstractUploadListener {
    * Default constructor.
    * 
    */
-  public UploadListener(int sleepMilliseconds, int requestSize) {
+  public UploadListener(int sleepMilliseconds, long requestSize) {
     super(sleepMilliseconds, requestSize);
     startWatcher();
   }
